Proposal: Allow for community reviewers to reduce PR backlog
I wanted to propose the following policy change in an effort to better
expidite our backlog of Pull Requests

Summary:
I'd like to create a github group within the openssl organization,
called reviewers, to which community members can nominate themselves to
join (and OTC can vote on), in an effort to create a larger pool of
people to review outstanding PR's in such a way that our merge
requirements overhead is reduced.

Details regarding the nomination process/abilities/requirements are in
the policy document

One of the things that has bothered me in reviewing our backlog is the
number of prs that "falls through the cracks".  We don't have the
resources available to review everything in a timely fashion, so I would
like to propose that we engage our community to give them additional
agency to do some of this work.

By allowing community members to nominate themselves to be reviewers, we
gain several advantages:

1) It allows for community members to aid in reducing the number of
   reviews people in the otc and committers groups need to do.  A
   reviewer group expands the number of people who's reviews "count"
   toward our minimal number to merge a request

2) It creates greater engagement and ownership in our communities.
   People who have the ability to affect change in our code base should
   translate into more people who want to participate